# docs/swagger.yaml
definitions:
...
  personalfinancedss_internal_module_cashflow_debt_domain.DebtStatus:
    enum:
    - active
    - paid_off
    - settled
    - defaulted
    - inactive
    type: string
    x-enum-comments:
      DebtStatusActive: Debt is active
      DebtStatusDefaulted: Debt is in default
      DebtStatusInactive: Debt is inactive
      DebtStatusPaidOff: Debt has been paid off
      DebtStatusSettled: Debt was settled
    x-enum-descriptions:
    - Debt is active
    - Debt has been paid off
    - Debt was settled
    - Debt is in default
    - Debt is inactive
    x-enum-varnames:
    - DebtStatusActive
    - DebtStatusPaidOff
    - DebtStatusSettled
    - DebtStatusDefaulted
    - DebtStatusInactive
  personalfinancedss_internal_module_cashflow_debt_domain.DebtType:
    enum:
    - credit_card
    - personal_loan
    - mortgage
    - auto_loan
    - student_loan
    - medical
    - other
    type: string
    x-enum-comments:
      DebtTypeAutoLoan: Auto loan
      DebtTypeCreditCard: Credit card debt
      DebtTypeMedical: Medical debt
      DebtTypeMortgage: Mortgage
      DebtTypeOther: Other debt
      DebtTypePersonalLoan: Personal loan
      DebtTypeStudentLoan: Student loan
    x-enum-descriptions:
    - Credit card debt
    - Personal loan
    - Mortgage
    - Auto loan
    - Student loan
    - Medical debt
    - Other debt
    x-enum-varnames:
    - DebtTypeCreditCard
    - DebtTypePersonalLoan
    - DebtTypeMortgage
    - DebtTypeAutoLoan
    - DebtTypeStudentLoan
    - DebtTypeMedical
    - DebtTypeOther
  personalfinancedss_internal_module_cashflow_debt_domain.PaymentFrequency:
    enum:
    - one_time
    - daily
    - weekly
    - biweekly
    - monthly
    - quarterly
    - yearly
    type: string
    x-enum-varnames:
    - FrequencyOneTime
    - FrequencyDaily
    - FrequencyWeekly
    - FrequencyBiweekly
    - FrequencyMonthly
    - FrequencyQuarterly
    - FrequencyYearly
  personalfinancedss_internal_module_cashflow_debt_dto.AddPaymentRequest:
    properties:
      amount:
        type: number
      date:
        type: string
      description:
        type: string
    required:
    - amount
    type: object
  personalfinancedss_internal_module_cashflow_debt_dto.CreateDebtRequest:
    properties:
      account_number:
        type: string
      creditor_name:
        type: string
      currency:
        type: string
      current_balance:
        type: number
      description:
        type: string
      due_date:
        type: string
      enable_reminders:
        type: boolean
      interest_rate:
        maximum: 100
        minimum: 0
        type: number
      linked_account_id:
        type: string
      minimum_payment:
        minimum: 0
        type: number
      name:
        type: string
      next_payment_date:
        type: string
      notes:
        type: string
      payment_amount:
        minimum: 0
        type: number
      payment_frequency:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.PaymentFrequency'
      principal_amount:
        type: number
      reminder_frequency:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.DebtStatus'
      tags:
        type: string
      type:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.DebtType'
    required:
    - currency
    - current_balance
    - name
    - principal_amount
    - start_date
    - type
    type: object
  personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse:
    properties:
      account_number:
        type: string
      created_at:
        type: string
      creditor_name:
        type: string
      currency:
        type: string
      current_balance:
        type: number
      days_until_next_payment:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      enable_reminders:
        type: boolean
      id:
        type: string
      interest_rate:
        type: number
      last_payment_amount:
        type: number
      last_payment_date:
        type: string
      last_reminder_sent_at:
        type: string
      linked_account_id:
        type: string
      minimum_payment:
        type: number
      name:
        type: string
      next_payment_date:
        type: string
      notes:
        type: string
      paid_off_date:
        type: string
      payment_amount:
        type: number
      payment_frequency:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.PaymentFrequency'
      percentage_paid:
        type: number
      principal_amount:
        type: number
      remaining_amount:
        type: number
      reminder_frequency:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.DebtStatus'
      tags:
        type: string
      total_interest_paid:
        type: number
      total_paid:
        type: number
      type:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.DebtType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  personalfinancedss_internal_module_cashflow_debt_dto.DebtSummaryResponse:
    properties:
      active_debts:
        type: integer
      average_progress:
        type: number
      debts_by_status:
        additionalProperties:
          type: integer
        type: object
      debts_by_type:
        additionalProperties:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_service.DebtTypeSum'
        type: object
      overdue_debts:
        type: integer
      paid_off_debts:
        type: integer
      total_current_balance:
        type: number
      total_debts:
        type: integer
      total_interest_paid:
        type: number
      total_paid:
        type: number
      total_principal_amount:
        type: number
      total_remaining:
        type: number
    type: object
  personalfinancedss_internal_module_cashflow_debt_dto.UpdateDebtRequest:
    properties:
      account_number:
        type: string
      creditor_name:
        type: string
      currency:
        type: string
      current_balance:
        minimum: 0
        type: number
      description:
        type: string
      due_date:
        type: string
      enable_reminders:
        type: boolean
      interest_rate:
        maximum: 100
        minimum: 0
        type: number
      last_payment_date:
        type: string
      linked_account_id:
        type: string
      minimum_payment:
        minimum: 0
        type: number
      name:
        type: string
      next_payment_date:
        type: string
      notes:
        type: string
      payment_amount:
        minimum: 0
        type: number
      payment_frequency:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.PaymentFrequency'
      principal_amount:
        type: number
      reminder_frequency:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.DebtStatus'
      tags:
        type: string
      type:
        $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_domain.DebtType'
    type: object
  personalfinancedss_internal_module_cashflow_debt_service.DebtTypeSum:
    properties:
      count:
        type: integer
      current_balance:
        type: number
      principal_amount:
        type: number
      progress:
        type: number
      total_paid:
        type: number
    type: object
...
info:
  contact:
    email: support@personalfinancedss.com
    name: API Support
    url: http://www.swagger.io/support
  description: Personal Finance Decision Support System API for managing accounts,
    transactions, budgets, goals, and investments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Personal Finance DSS API
  version: "1.0"
paths:
...
  /api/v1/debts:
    get:
      description: Get all debts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all user debts
      tags:
      - debts
    post:
      consumes:
      - application/json
      description: Create a new debt for the authenticated user
      parameters:
      - description: Debt details
        in: body
        name: debt
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.CreateDebtRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new debt
      tags:
      - debts
  /api/v1/debts/{id}:
    delete:
      description: Delete a debt
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete debt
      tags:
      - debts
    get:
      description: Get a specific debt by its ID
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get debt by ID
      tags:
      - debts
    put:
      consumes:
      - application/json
      description: Update an existing debt
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      - description: Debt details
        in: body
        name: debt
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.UpdateDebtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update debt
      tags:
      - debts
  /api/v1/debts/{id}/paid-off:
    post:
      description: Mark a debt as paid off
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Mark debt as paid off
      tags:
      - debts
  /api/v1/debts/{id}/payment:
    post:
      consumes:
      - application/json
      description: Add a payment amount to a debt
      parameters:
      - description: Debt ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.AddPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Add payment to debt
      tags:
      - debts
  /api/v1/debts/active:
    get:
      description: Get all active debts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get active debts
      tags:
      - debts
  /api/v1/debts/paid-off:
    get:
      description: Get all paid off debts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get paid off debts
      tags:
      - debts
  /api/v1/debts/summary:
    get:
      description: Get a summary of all debts for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_module_cashflow_debt_dto.DebtSummaryResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get debt summary
      tags:
      - debts
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme. Enter 'Bearer'
      [space] and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"